+++ common/src/main/java/edu/snu/cay/common/math/linalg/breeze/DenseMatrix.java
@@ -27,13 +27,13 @@ import scala.runtime.RichInt;
  */
 public class DenseMatrix implements Matrix {
 
-  private final breeze.linalg.DenseMatrix<Double> breezeMatrix;
+  private final breeze.linalg.DenseMatrix<Float> breezeMatrix;
 
-  DenseMatrix(final breeze.linalg.DenseMatrix<Double> breezeMatrix) {
+  DenseMatrix(final breeze.linalg.DenseMatrix<Float> breezeMatrix) {
     this.breezeMatrix = breezeMatrix;
   }
 
-  breeze.linalg.DenseMatrix<Double> getBreezeMatrix() {
+  breeze.linalg.DenseMatrix<Float> getBreezeMatrix() {
     return breezeMatrix;
   }
 
@@ -89,7 +89,7 @@ public class DenseMatrix implements Matrix {
    * @return element specified by given indices
    */
   @Override
-  public double get(final int rowIndex, final int columnIndex) {
+  public Float get(final int rowIndex, final int columnIndex) {
     return breezeMatrix.apply(rowIndex, columnIndex);
   }
 
@@ -98,7 +98,7 @@ public class DenseMatrix implements Matrix {
    */
   @Override
   public DenseVector sliceColumn(final int index) {
-    return new DenseVector((breeze.linalg.DenseVector<Double>)
+    return new DenseVector((breeze.linalg.DenseVector<Float>)
         breezeMatrix.apply(MatrixOps.COLON_COLON, index, MatrixOps.SLICE_COL_D));
   }
 
@@ -107,7 +107,7 @@ public class DenseMatrix implements Matrix {
    */
   @Override
   public DenseVector sliceRow(final int index) {
-    return new DenseVector((breeze.linalg.DenseVector<Double>) ((breeze.linalg.Transpose)
+    return new DenseVector((breeze.linalg.DenseVector<Float>) ((breeze.linalg.Transpose)
         breezeMatrix.apply(index, MatrixOps.COLON_COLON, MatrixOps.SLICE_ROW_D)).inner());
   }
 
@@ -116,7 +116,7 @@ public class DenseMatrix implements Matrix {
    */
   @Override
   public DenseMatrix sliceColumns(final int start, final int end) {
-    return new DenseMatrix((breeze.linalg.DenseMatrix<Double>)
+    return new DenseMatrix((breeze.linalg.DenseMatrix<Float>)
         breezeMatrix.apply(MatrixOps.COLON_COLON, new RichInt(start).until(end), MatrixOps.SLICE_COLS_D));
   }
 
@@ -125,7 +125,7 @@ public class DenseMatrix implements Matrix {
    */
   @Override
   public DenseMatrix sliceRows(final int start, final int end) {
-    return new DenseMatrix((breeze.linalg.DenseMatrix<Double>)
+    return new DenseMatrix((breeze.linalg.DenseMatrix<Float>)
         breezeMatrix.apply(new RichInt(start).until(end), MatrixOps.COLON_COLON, MatrixOps.SLICE_ROWS_D));
   }
 
@@ -194,7 +194,7 @@ public class DenseMatrix implements Matrix {
    * @param value given value
    */
   @Override
-  public void set(final int rowIndex, final int columnIndex, final double value) {
+  public void set(final int rowIndex, final int columnIndex, final Float value) {
     breezeMatrix.update(rowIndex, columnIndex, value);
   }
 
@@ -204,7 +204,7 @@ public class DenseMatrix implements Matrix {
    */
   @Override
   public Matrix transpose() {
-    return new DenseMatrix((breeze.linalg.DenseMatrix<Double>) breezeMatrix.t(MatrixOps.T_D));
+    return new DenseMatrix((breeze.linalg.DenseMatrix<Float>) breezeMatrix.t(MatrixOps.T_D));
   }
 
   /**
@@ -219,8 +219,8 @@ public class DenseMatrix implements Matrix {
    * Converts this matrix to a flat array (column-major).
    * @return flat array with matrix elements
    */
-  public double[] toArray() {
-    return (double[]) breezeMatrix.toArray();
+  public Float[] toArray() {
+    return (Float[]) (breezeMatrix.toArray());
   }
 
   /**
@@ -261,8 +261,8 @@ public class DenseMatrix implements Matrix {
    * @return new {@link DenseMatrix} with operation result
    */
   @Override
-  public Matrix add(final double value) {
-    return new DenseMatrix((breeze.linalg.DenseMatrix<Double>) breezeMatrix.$plus(value, MatrixOps.ADD_DT));
+  public Matrix add(final Float value) {
+    return new DenseMatrix((breeze.linalg.DenseMatrix<Float>) breezeMatrix.$plus(value, MatrixOps.ADD_DT));
   }
 
   /**
@@ -271,7 +271,7 @@ public class DenseMatrix implements Matrix {
    * @return this matrix with operation result
    */
   @Override
-  public Matrix addi(final double value) {
+  public Matrix addi(final Float value) {
     ((NumericOps) breezeMatrix).$plus$eq(value, MatrixOps.ADDI_DT);
     return this;
   }
@@ -284,10 +284,10 @@ public class DenseMatrix implements Matrix {
   @Override
   public Matrix add(final Matrix matrix) {
     if (matrix.isDense()) {
-      return new DenseMatrix((breeze.linalg.DenseMatrix<Double>)
+      return new DenseMatrix((breeze.linalg.DenseMatrix<Float>)
           breezeMatrix.$plus(((DenseMatrix) matrix).breezeMatrix, MatrixOps.ADD_DD));
     } else {
-      return new DenseMatrix((breeze.linalg.DenseMatrix<Double>)
+      return new DenseMatrix((breeze.linalg.DenseMatrix<Float>)
           breezeMatrix.$plus(((CSCMatrix) matrix).getBreezeMatrix(), MatrixOps.ADD_DS));
     }
   }
@@ -313,8 +313,8 @@ public class DenseMatrix implements Matrix {
    * @return new {@link DenseMatrix} with operation result
    */
   @Override
-  public Matrix sub(final double value) {
-    return new DenseMatrix((breeze.linalg.DenseMatrix<Double>) breezeMatrix.$minus(value, MatrixOps.SUB_DT));
+  public Matrix sub(final Float value) {
+    return new DenseMatrix((breeze.linalg.DenseMatrix<Float>) breezeMatrix.$minus(value, MatrixOps.SUB_DT));
   }
 
   /**
@@ -323,7 +323,7 @@ public class DenseMatrix implements Matrix {
    * @return this matrix with operation result
    */
   @Override
-  public Matrix subi(final double value) {
+  public Matrix subi(final Float value) {
     ((NumericOps) breezeMatrix).$minus$eq(value, MatrixOps.SUBI_DT);
     return this;
   }
@@ -336,10 +336,10 @@ public class DenseMatrix implements Matrix {
   @Override
   public Matrix sub(final Matrix matrix) {
     if (matrix.isDense()) {
-      return new DenseMatrix((breeze.linalg.DenseMatrix<Double>)
+      return new DenseMatrix((breeze.linalg.DenseMatrix<Float>)
           breezeMatrix.$minus(((DenseMatrix) matrix).breezeMatrix, MatrixOps.SUB_DD));
     } else {
-      return new DenseMatrix((breeze.linalg.DenseMatrix<Double>)
+      return new DenseMatrix((breeze.linalg.DenseMatrix<Float>)
           breezeMatrix.$minus(((CSCMatrix) matrix).getBreezeMatrix(), MatrixOps.SUB_DS));
     }
   }
@@ -365,8 +365,8 @@ public class DenseMatrix implements Matrix {
    * @return new {@link DenseMatrix} with operation result
    */
   @Override
-  public Matrix mul(final double value) {
-    return new DenseMatrix((breeze.linalg.DenseMatrix<Double>) breezeMatrix.$colon$times(value, MatrixOps.MUL_DT));
+  public Matrix mul(final Float value) {
+    return new DenseMatrix((breeze.linalg.DenseMatrix<Float>) breezeMatrix.$colon$times(value, MatrixOps.MUL_DT));
   }
 
   /**
@@ -375,7 +375,7 @@ public class DenseMatrix implements Matrix {
    * @return this matrix with operation result
    */
   @Override
-  public Matrix muli(final double value) {
+  public Matrix muli(final Float value) {
     ((NumericOps) breezeMatrix).$colon$times$eq(value, MatrixOps.MULI_DT);
     return this;
   }
@@ -388,10 +388,10 @@ public class DenseMatrix implements Matrix {
   @Override
   public Matrix mul(final Matrix matrix) {
     if (matrix.isDense()) {
-      return new DenseMatrix((breeze.linalg.DenseMatrix<Double>)
+      return new DenseMatrix((breeze.linalg.DenseMatrix<Float>)
           breezeMatrix.$colon$times(((DenseMatrix) matrix).breezeMatrix, MatrixOps.EMUL_DD));
     } else {
-      return new DenseMatrix((breeze.linalg.DenseMatrix<Double>)
+      return new DenseMatrix((breeze.linalg.DenseMatrix<Float>)
           breezeMatrix.$colon$times(((CSCMatrix) matrix).getBreezeMatrix(), MatrixOps.EMUL_MM));
     }
   }
@@ -417,8 +417,8 @@ public class DenseMatrix implements Matrix {
    * @return new {@link DenseMatrix} with operation result
    */
   @Override
-  public Matrix div(final double value) {
-    return new DenseMatrix((breeze.linalg.DenseMatrix<Double>) breezeMatrix.$div(value, MatrixOps.DIV_DT));
+  public Matrix div(final Float value) {
+    return new DenseMatrix((breeze.linalg.DenseMatrix<Float>) breezeMatrix.$div(value, MatrixOps.DIV_DT));
   }
 
   /**
@@ -427,7 +427,7 @@ public class DenseMatrix implements Matrix {
    * @return this matrix with operation result
    */
   @Override
-  public Matrix divi(final double value) {
+  public Matrix divi(final Float value) {
     ((NumericOps) breezeMatrix).$div$eq(value, MatrixOps.DIVI_DT);
     return this;
   }
@@ -440,10 +440,10 @@ public class DenseMatrix implements Matrix {
   @Override
   public Matrix div(final Matrix matrix) {
     if (matrix.isDense()) {
-      return new DenseMatrix((breeze.linalg.DenseMatrix<Double>)
+      return new DenseMatrix((breeze.linalg.DenseMatrix<Float>)
           breezeMatrix.$div(((DenseMatrix) matrix).breezeMatrix, MatrixOps.EDIV_DD));
     } else {
-      return new DenseMatrix((breeze.linalg.DenseMatrix<Double>)
+      return new DenseMatrix((breeze.linalg.DenseMatrix<Float>)
           breezeMatrix.$div(((CSCMatrix) matrix).getBreezeMatrix(), MatrixOps.EDIV_MM));
     }
   }
@@ -471,10 +471,10 @@ public class DenseMatrix implements Matrix {
   @Override
   public Vector mmul(final Vector vector) {
     if (vector.isDense()) {
-      return new DenseVector((breeze.linalg.DenseVector<Double>)
+      return new DenseVector((breeze.linalg.DenseVector<Float>)
           breezeMatrix.$times(((DenseVector) vector).getBreezeVector(), MatrixOps.MUL_DMDV));
     } else {
-      return new DenseVector((breeze.linalg.DenseVector<Double>)
+      return new DenseVector((breeze.linalg.DenseVector<Float>)
           breezeMatrix.$times(((SparseVector) vector).getBreezeVector(), MatrixOps.MUL_DMSV));
     }
   }
@@ -487,10 +487,10 @@ public class DenseMatrix implements Matrix {
   @Override
   public Matrix mmul(final Matrix matrix) {
     if (matrix.isDense()) {
-      return new DenseMatrix((breeze.linalg.DenseMatrix<Double>)
+      return new DenseMatrix((breeze.linalg.DenseMatrix<Float>)
           breezeMatrix.$times(((DenseMatrix) matrix).breezeMatrix, MatrixOps.MUL_DMDM));
     } else {
-      return new DenseMatrix((breeze.linalg.DenseMatrix<Double>)
+      return new DenseMatrix((breeze.linalg.DenseMatrix<Float>)
           breezeMatrix.$times(((CSCMatrix) matrix).getBreezeMatrix(), MatrixOps.MUL_DMSM));
     }
   }
